requirements.txt
cryptography>=41.0
pyyaml>=6.0
pytest>=8.0
fastapi>=0.109.0
uvicorn>=0.27.0
